Consolidate most Remote/Deployment functionality under Client class
* Deployment class has been removed
* Config class handles Client creation
* RemoteVPS functionality now handled by Client
