+++ client/client_test.go
@@ -1,25 +1,26 @@
 package client
 
 import (
+	"crypto/tls"
 	"encoding/json"
+	"fmt"
 	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
+	"os"
 	"strings"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/ubclaunchpad/inertia/common"
-	git "gopkg.in/src-d/go-git.v4"
-	"gopkg.in/src-d/go-git.v4/config"
-	"gopkg.in/src-d/go-git.v4/storage/memory"
 )
 
 var (
 	fakeAuth = "ubclaunchpad"
 )
 
-func getMockDeployment(ts *httptest.Server, s *memory.Storage) (*Deployment, error) {
+func getMockClient(ts *httptest.Server) *Client {
 	var (
 		url  string
 		port string
@@ -40,28 +41,131 @@ func getMockDeployment(ts *httptest.Server, s *memory.Storage) (*Deployment, err
 		Daemon: &DaemonConfig{
 			Port:   port,
 			Secret: "arjan",
+			Token:  fakeAuth,
 		},
 	}
-	mockRepo, err := git.Init(s, nil)
-	if err != nil {
-		return nil, err
-	}
-	_, err = mockRepo.CreateRemote(&config.RemoteConfig{
-		Name: "origin",
-		URLs: []string{"myremote"},
-	})
-	if err != nil {
-		return nil, err
-	}
 
-	return &Deployment{
-		RemoteVPS:  mockRemote,
-		Repository: mockRepo,
-		Auth:       fakeAuth,
-		Project:    "test_project",
-	}, nil
+	return &Client{
+		RemoteVPS: mockRemote,
+		project:   "test_project",
+	}
 }
 
+func getIntegrationClient() *Client {
+	remote := &RemoteVPS{
+		IP:   "127.0.0.1",
+		PEM:  "../test/keys/id_rsa",
+		User: "root",
+		Daemon: &DaemonConfig{
+			Port: "4303",
+		},
+	}
+	travis := os.Getenv("TRAVIS")
+	if travis != "" {
+		remote.Daemon.SSHPort = "69"
+	} else {
+		remote.Daemon.SSHPort = "22"
+	}
+	return &Client{
+		RemoteVPS: remote,
+	}
+}
+
+func TestInstallDocker(t *testing.T) {
+	remote := getIntegrationClient()
+	script, err := ioutil.ReadFile("bootstrap/docker.sh")
+	assert.Nil(t, err)
+
+	// Make sure the right command is run.
+	session := mockSSHRunner{c: remote}
+	remote.installDocker(&session)
+	assert.Equal(t, string(script), session.Calls[0])
+}
+
+func TestDaemonUp(t *testing.T) {
+	remote := getIntegrationClient()
+	script, err := ioutil.ReadFile("bootstrap/daemon-up.sh")
+	assert.Nil(t, err)
+	actualCommand := fmt.Sprintf(string(script), "latest", "4303", "0.0.0.0")
+
+	// Make sure the right command is run.
+	session := mockSSHRunner{c: remote}
+
+	// Make sure the right command is run.
+	err = remote.DaemonUp(&session, "latest", "0.0.0.0", "4303")
+	assert.Nil(t, err)
+	println(actualCommand)
+	assert.Equal(t, actualCommand, session.Calls[0])
+}
+
+func TestKeyGen(t *testing.T) {
+	remote := getIntegrationClient()
+	script, err := ioutil.ReadFile("bootstrap/token.sh")
+	assert.Nil(t, err)
+	tokenScript := fmt.Sprintf(string(script), "test")
+
+	// Make sure the right command is run.
+	session := mockSSHRunner{c: remote}
+
+	// Make sure the right command is run.
+	_, err = remote.getDaemonAPIToken(&session, "test")
+	assert.Nil(t, err)
+	assert.Equal(t, session.Calls[0], tokenScript)
+}
+
+func TestBootstrap(t *testing.T) {
+	remote := getIntegrationClient()
+	dockerScript, err := ioutil.ReadFile("bootstrap/docker.sh")
+	assert.Nil(t, err)
+
+	keyScript, err := ioutil.ReadFile("bootstrap/keygen.sh")
+	assert.Nil(t, err)
+
+	script, err := ioutil.ReadFile("bootstrap/token.sh")
+	assert.Nil(t, err)
+	tokenScript := fmt.Sprintf(string(script), "test")
+
+	script, err = ioutil.ReadFile("bootstrap/daemon-up.sh")
+	assert.Nil(t, err)
+	daemonScript := fmt.Sprintf(string(script), "test", "4303", "127.0.0.1")
+
+	session := mockSSHRunner{c: remote}
+	err = remote.BootstrapRemote(&session)
+	assert.Nil(t, err)
+
+	// Make sure all commands are formatted correctly
+	assert.Equal(t, string(dockerScript), session.Calls[0])
+	assert.Equal(t, string(keyScript), session.Calls[1])
+	assert.Equal(t, daemonScript, session.Calls[2])
+	assert.Equal(t, tokenScript, session.Calls[3])
+}
+
+func TestBootstrapIntegration(t *testing.T) {
+	if testing.Short() {
+		t.Skip("skipping integration test")
+	}
+
+	remote := getIntegrationClient()
+	session := &SSHRunner{c: remote}
+	err := remote.BootstrapRemote(session)
+	assert.Nil(t, err)
+
+	// Daemon setup takes a bit of time - do a crude wait
+	time.Sleep(3 * time.Second)
+
+	// Check if daemon is online following bootstrap
+	host := "https://" + remote.GetIPAndPort()
+	tr := &http.Transport{
+		TLSClientConfig: &tls.Config{InsecureSkipVerify: true},
+	}
+	client := &http.Client{Transport: tr}
+	resp, err := client.Get(host)
+	assert.Nil(t, err)
+	assert.Equal(t, resp.StatusCode, http.StatusOK)
+	defer resp.Body.Close()
+	_, err = ioutil.ReadAll(resp.Body)
+	assert.Nil(t, err)
+}
 func TestUp(t *testing.T) {
 	testServer := httptest.NewTLSServer(http.HandlerFunc(func(rw http.ResponseWriter, req *http.Request) {
 		rw.WriteHeader(http.StatusOK)
@@ -90,13 +194,8 @@ func TestUp(t *testing.T) {
 	}))
 	defer testServer.Close()
 
-	memory := memory.NewStorage()
-	defer func() { memory = nil }()
-
-	d, err := getMockDeployment(testServer, memory)
-	assert.Nil(t, err)
-
-	resp, err := d.Up("docker-compose", false)
+	d := getMockClient(testServer)
+	resp, err := d.Up("myremote.git", "docker-compose", false)
 	assert.Nil(t, err)
 	assert.Equal(t, http.StatusOK, resp.StatusCode)
 }
@@ -117,12 +216,7 @@ func TestDown(t *testing.T) {
 	}))
 	defer testServer.Close()
 
-	memory := memory.NewStorage()
-	defer func() { memory = nil }()
-
-	d, err := getMockDeployment(testServer, memory)
-	assert.Nil(t, err)
-
+	d := getMockClient(testServer)
 	resp, err := d.Down()
 	assert.Nil(t, err)
 	assert.Equal(t, http.StatusOK, resp.StatusCode)
@@ -144,25 +238,15 @@ func TestStatus(t *testing.T) {
 	}))
 	defer testServer.Close()
 
-	memory := memory.NewStorage()
-	defer func() { memory = nil }()
-
-	d, err := getMockDeployment(testServer, memory)
-	assert.Nil(t, err)
-
+	d := getMockClient(testServer)
 	resp, err := d.Status()
 	assert.Nil(t, err)
 	assert.Equal(t, http.StatusOK, resp.StatusCode)
 }
 
 func TestStatusFail(t *testing.T) {
-	memory := memory.NewStorage()
-	defer func() { memory = nil }()
-
-	d, err := getMockDeployment(nil, memory)
-	assert.Nil(t, err)
-
-	_, err = d.Status()
+	d := getMockClient(nil)
+	_, err := d.Status()
 	assert.Contains(t, err.Error(), "appears offline")
 }
 
@@ -182,12 +266,7 @@ func TestReset(t *testing.T) {
 	}))
 	defer testServer.Close()
 
-	memory := memory.NewStorage()
-	defer func() { memory = nil }()
-
-	d, err := getMockDeployment(testServer, memory)
-	assert.Nil(t, err)
-
+	d := getMockClient(testServer)
 	resp, err := d.Reset()
 	assert.Nil(t, err)
 	assert.Equal(t, http.StatusOK, resp.StatusCode)
@@ -219,12 +298,7 @@ func TestLogs(t *testing.T) {
 	}))
 	defer testServer.Close()
 
-	memory := memory.NewStorage()
-	defer func() { memory = nil }()
-
-	d, err := getMockDeployment(testServer, memory)
-	assert.Nil(t, err)
-
+	d := getMockClient(testServer)
 	resp, err := d.Logs(true, "docker-compose")
 	assert.Nil(t, err)
 	assert.Equal(t, http.StatusOK, resp.StatusCode)
